doc: update examples, now implment Resolver decorator
